Functionally working
Web app is successfully producing a randomly generated password based on user input. User can determine quantity of characters and whether or not to use special characters. Next, I need to work on the UI. As of now UI has been extremely basic just to work on implementing the functionality properly.